requirements.txt
fastapi
uvicorn
pandas
numpy
google-genai
python-dotenv
